Separate moveThought integration tests from reducer tests.
